Add functionality to create gist from the editor
This commit adds functionality to create a gist right from the editor
using the already existing underlying logic. However, the current
behaviour of the 'create gist' button at the left of status bar remains
the same.

FEATURES ADDED
- Create Gist of selected text from inside a file
- Create Gist of a whole file using a button in the titlebar tabs view
- Create Gist of a file in the explorer
- Create Gist of all files in a folder in the explorer view (NOTE: this
  recursively loops through every file and sub-files in the cuurent
  folder and sub folders)

HANDLED ERRORS
When triggering the commands from the command pallete in vscode, there
might be some error, e.g
- 'Create gist from active editor' only works when there is an active
  editor
- 'Create fist from explorer' only works when there a file or folder is
  selected from the explorer
- 'Create gist from selection' only works when text is selected inside a
  an editor
All the errors were handled, but might need to be added to the
documentation

Overall this commit comes with a little refactor, but the underlying
logic remains untouched.